requirements.txt
absl-py==1.4.0
altair==4.2.2
astor==0.8.1
astunparse==1.6.3
attrs==22.2.0
base58==2.1.1
beautifulsoup4==4.12.2
blinker==1.6.1
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
entrypoints==0.4
Faker==18.11.1
favicon==0.7.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
htbuilder==0.6.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
more-itertools==9.1.0
murmurhash==1.0.9
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
pandas==1.5.3
pathy==0.10.2
Pillow==9.4.0
preshed==3.0.8
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.9
pydeck==0.8.0
pymdown-extensions==10.0.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
st-annotated-text==4.0.0
streamlit==1.21.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.5
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-gallery==0.1.0
streamlit-image-coordinates==0.1.5
streamlit-keyup==0.2.0
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow_intel==2.10.0
termcolor==2.2.0
thinc==8.1.10
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tqdm==4.65.0
typer==0.7.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
Werkzeug==2.2.2
wrapt==1.14.1
